Only take gold when game is not lost yet.
This fixes the loseCrossing failure. The structure required that there
was gold at (1,3) at all times, since the game would be over as soon as
Pacman crossed the ghost and thus could no longer pick up any gold.

However, we never checked for game over in our gold pick up rule, which
meant that we still picked up the gold. Of course, this meant that we
would never find a model as the gold would be gone at t=2. By checking
for ~GameLost(t) before removing the gold, we take this into account.